+++ src/main/java/com/zerobase/CafeBom/CafeBomApplication.java
@@ -1,4 +1,4 @@
-package com.zerobase.CafeBom;
+package com.zerobase.cafebom;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
