+++ src/main/java/com/zerobase/CafeBom/user/security/JwtAuthenticationFilter.java
@@ -1,4 +1,4 @@
-package com.zerobase.CafeBom.config;
+package com.zerobase.cafebom.user.security;
 
 import java.io.IOException;
 import javax.servlet.FilterChain;
